# Backend - nginx/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Реализация RestAPI на Go с использование Swagger для документации",
        "title": "School 21 ApiFromSpec docs",
        "termsOfService": "http://swagger.io/terms",
        "contact": {
            "name": "desmonsu"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/products": {
            "get": {
                "description": "Возвращает массив товаров. Если товары не найдены, возвращает пустой массив.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получение всех товаров",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/apifromspec_internal_pkg.GetProductResponse"
                            }
                        }
                    },
                    "500": {
                        "description": "fail to fetch products",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "description": "Возвращает информацию об успешности операции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Добавление товара в базу данных",
                "parameters": [
                    {
                        "description": "Данные для добавления товара",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.AddProductRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "product added successfully",
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_domain.Product"
                        }
                    },
                    "400": {
                        "description": "wrong input format or validation error",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to create product",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/products/{id}": {
            "get": {
                "description": "Возвращает найденный товар",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Получение товара по id",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.GetProductResponse"
                        }
                    },
                    "400": {
                        "description": "wrong input format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "product not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Возвращает информацию об успешности операции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Удаляет товар из базы данных",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID товара для удаления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "product deleted successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "wrong input format or can not parse id from URL",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "product not found",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "failed to delete product",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/products/{id}/subtract": {
            "patch": {
                "description": "Возвращает информацию об успешности операции",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "products"
                ],
                "summary": "Вычитание товара из базы при наличии",
                "parameters": [
                    {
                        "type": "string",
                        "format": "uuid",
                        "description": "ID товара",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Данные для вычитания товара",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/apifromspec_internal_pkg.SubtractProductRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "amount decreased successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "wrong input format, validation error or insufficient stock",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apifromspec_internal_domain.Product": {
            "type": "object",
            "properties": {
                "availableStock": {
                    "type": "integer"
                },
                "category": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "imageID": {
                    "type": "string"
                },
                "lastUpdateDate": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "integer",
                    "format": "int64"
                },
                "supplierID": {
                    "type": "string"
                }
            }
        },
...
        "apifromspec_internal_pkg.AddProductRequest": {
            "type": "object",
            "properties": {
                "available_stock": {
                    "type": "integer",
                    "example": 10
                },
                "category": {
                    "type": "string",
                    "example": "Laptops"
                },
                "image": {
                    "type": "string",
                    "format": "base64"
                },
                "name": {
                    "type": "string",
                    "example": "Laptop Air"
                },
                "price": {
                    "type": "integer",
                    "example": 120000
                },
                "supplier_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "apifromspec_internal_pkg.GetProductResponse": {
            "type": "object",
            "properties": {
                "available_stock": {
                    "type": "integer",
                    "example": 10
                },
                "category": {
                    "type": "string",
                    "example": "Laptops"
                },
                "id": {
                    "type": "string",
                    "example": "770e8400-e29b-41d4-a716-446655440022"
                },
                "image_id": {
                    "type": "string",
                    "example": "880e8400-e29b-41d4-a716-446655440033"
                },
                "last_update_date": {
                    "type": "string",
                    "example": "2026-04-28T12:00:00Z"
                },
                "name": {
                    "type": "string",
                    "example": "Laptop Air"
                },
                "price": {
                    "type": "integer",
                    "example": 120000
                },
                "supplier_id": {
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
...
        "apifromspec_internal_pkg.SubtractProductRequest": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer",
                    "example": 3
                }
            }
        }
    }
}